refactor(desktop): 提取 host service 纯工具与内部上下文
